requirements.txt
llama-cpp-python ==0.2.11
pypdf ==3.16.4
python-docx ==0.8.11
PySide6 ==6.6.1
sentence-transformers ==2.2.2
faiss-cpu ==1.7.4
langchain ==0.0.334
InstructorEmbedding==1.0.1
huggingface-hub==0.11.1
transformers==4.25.1
tokenizers==0.13.2